Define the fatalError() method in the hacky eval()'d pseudo-class in a way that does *not* exit(), like how Maintenance#fatalError does
This gets at least deleteBatch working again and that's good enough for me.
The other changes are probably not needed but maybe/hopefully good code hygiene,
given that the SpecialPage object representing the Special:Maintenance page is not going away
in the future, but the $wgOut global object on the other hand *is*.

Tested locally on my 1.35 box by deleting two pages with deleteBatch.

Caveats:
* core deleteBatch.php maintenance script defaults to "Delete page script" if no username is chosen; this behavior for some reason is not replicated here...
** and if you manually try to supply "Delete page script" as the username to use, it _will_ fail: CannotCreateActorException from line 285 of /var/www/html/135/includes/ActorMigration.php: Cannot create an actor for a usable name that is not an existing user: user_id=0 user_name="Delete page script"
** but deleting pages as "MediaWiki default" or an existing user account certainly seemed to work just as intended

Change-Id: I2d2e14a5f2bb180a3910d4188bd467553fe48756